Configure manifest attrs for Kotlin JVM target's artifactsTask
