+++ src/main/kotlin/com/github/jengelman/gradle/plugins/shadow/internal/GradleCompat.kt
@@ -15,9 +15,6 @@ import org.gradle.api.provider.Property
 import org.gradle.api.provider.Provider
 import org.gradle.api.provider.SetProperty
 import org.gradle.api.tasks.SourceSetContainer
-import org.gradle.api.tasks.TaskContainer
-import org.gradle.api.tasks.TaskProvider
-import org.gradle.jvm.tasks.Jar
 import org.gradle.jvm.toolchain.JavaToolchainService
 
 /**
@@ -55,9 +52,6 @@ internal fun Project.addBuildScanCustomValues() {
     }
   }
 }
-
-internal inline val TaskContainer.jar: TaskProvider<Jar>
-  get() = named("jar", Jar::class.java)
 
 internal inline fun <reified V : Any> ObjectFactory.property(
   defaultValue: Any? = null,
